pyproject.toml
[project]
name = "passport-photo-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.4",
    "numpy>=2.2.6",
    "onnxruntime>=1.23.2",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "rembg>=2.0.69",
    "uvicorn>=0.38.0",
]
